Extract async scheduling methods to AsyncRequestScheduling
